+++ homeassistant/components/advantage_air/binary_sensor.py
@@ -1,6 +1,5 @@
 """Binary Sensor platform for Advantage Air integration."""
 
-from homeassistant.components.advantage_air import AdvantageAirEntity
 from homeassistant.components.binary_sensor import (
     DEVICE_CLASS_MOTION,
     DEVICE_CLASS_PROBLEM,
@@ -8,6 +7,7 @@ from homeassistant.components.binary_sensor import (
 )
 
 from .const import DOMAIN as ADVANTAGE_AIR_DOMAIN
+from .entity import AdvantageAirEntity
 
 
 async def async_setup_entry(hass, config_entry, async_add_entities):
